IHRAA-17731 - Attempt to fix PR checks and general housekeeping.
