+++ src/main/java/com/iheartradio/android/modules/crowdcontrol/SendOverHTTP.java
@@ -1,4 +1,4 @@
-package com.lotame.android;
+package com.iheartradio.android.modules.crowdcontrol;
 
 import android.os.AsyncTask;
 import android.util.Log;
